+++ CHANGELOG.md
@@ -6,7 +6,14 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased][unreleased]
 *Nothing So Far*
-### [0.6.7] - 2015-12-02
+
+### [0.7.0] - 2015-12-10
+### Changed
+- Update travis build for evm 0.7 + template
+### Added
+- add command to update travis config
+
+## [0.6.7] - 2015-12-02
 ### Changed
 - Wrong syntax now trigger an error
 ## [0.6.6] - 2015-11-18
@@ -17,6 +24,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## More history digging would be done one day
 
-[unreleased]: https://github.com/AdrieanKhisbe/cask-package-toolset.el/compare/v0.6.7...HEAD
+[unreleased]: https://github.com/AdrieanKhisbe/cask-package-toolset.el/compare/v0.7.0...HEAD
+[unreleased]: https://github.com/AdrieanKhisbe/cask-package-toolset.el/compare/v0.6.7...v0.7.0
 [0.6.7]: https://github.com/AdrieanKhisbe/cas-package-toolset.el/compare/v0.6.6...v0.6.7
 [0.6.6]: https://github.com/AdrieanKhisbe/cas-package-toolset.el/compare/v0.6.5...v0.6.6